refactor: extract request routing to dispatcher module
Rollback trigger: Rollback if there are any regressions in VRAM residency polling or NBD connectivity.
