+++ crates/ramshared-wsl2d/src/broker_srv/dispatcher.rs
@@ -1,0 +1,1483 @@
+
+use std::time::Instant;
+
+use ramshared_broker::arbiter::Action;
+use ramshared_broker::lease::LeaseDeny;
+use ramshared_broker::model::{PsiSample, Slice, SliceId, SliceState, TenantId, TransportKind};
+use ramshared_broker::protocol::{Msg, NbdEndpoint, PROTO_VERSION, SwapEntry, TenantMem, TenantStatus};
+
+use super::{BrokerCore, TenantState};
+use crate::telemetry::{ReconcileFlag, TelemetryCore, reconcile, ReconcileInput, vram_outros};
+use crate::canary_probe::CANARY_BYTES;
+use ramshared_broker::lease::LeaseDecision;
+use ramshared_broker::protocol::SliceIo;
+use std::sync::atomic::Ordering;
+use ramshared_broker::arbiter::TenantView;
+
+
+/// NBD endpoints that agents receive in `SwapOn`, chosen by the transport (DT-25).
+#[derive(Clone, Debug, Default)]
+pub struct EndpointCfg {
+    pub nbd_unix: Option<String>,
+    pub nbd_tcp: Option<(String, u16)>,
+}
+
+
+/// Input event for the core (produced by the IO layer). `sid` = connection ID (session).
+#[derive(Clone, Debug, PartialEq)]
+pub enum CoreEvent {
+    Msg(usize, Msg),
+    Disconnected(usize),
+    ZeroDone(SliceId, bool),
+    Demote(String),
+    Tick,
+}
+
+/// Output action executed by the IO layer.
+#[derive(Clone, Debug, PartialEq)]
+pub enum Outbound {
+    ToSession(usize, Msg),
+    CloseSession(usize),
+    ZeroSlice {
+        slice: SliceId,
+        base: u64,
+        len: u64,
+    },
+    Log(String),
+    /// Reconciled telemetry sample (RF-5); the IO layer timestamps `t`/`branch`/`commit`.
+    Telemetry(TelemetryCore),
+}
+
+
+/// Grace period (in ticks) before the 1st zero retry — gives time for `ZeroDone` in flight. Above this, the
+/// tick re-emits `ZeroExport`; in [`ZERO_RETRY_ERROR`] ticks without confirmation, logs ERROR (R4).
+const ZERO_RETRY_GRACE: u32 = 1;
+const ZERO_RETRY_ERROR: u32 = 5;
+
+/// Maps only an exact NBD block-device identity to its slice (DT-10/DT-21).
+fn dev_to_slice(dev: &str) -> Option<SliceId> {
+    let dev = dev.trim();
+    let base = if let Some(base) = dev.strip_prefix("/dev/") {
+        base
+    } else if let Some(base) = dev.strip_prefix('/') {
+        base
+    } else if !dev.contains('/') {
+        dev
+    } else {
+        return None;
+    };
+    let slice = base.strip_prefix("nbd")?;
+    if slice.is_empty() || !slice.bytes().all(|byte| byte.is_ascii_digit()) {
+        return None;
+    }
+    slice.parse().ok()
+}
+
+impl BrokerCore {
+    pub fn handle(&mut self, ev: CoreEvent, now: Instant) -> Vec<Outbound> {
+        let mut out = Vec::new();
+        match ev {
+            CoreEvent::Msg(sid, msg) => self.on_msg(sid, msg, &mut out),
+            CoreEvent::Disconnected(sid) => self.on_disconnect(sid, &mut out),
+            CoreEvent::ZeroDone(slice, ok) => self.on_zero_done(slice, ok, &mut out),
+            CoreEvent::Demote(reason) => self.on_demote(&reason, &mut out),
+            CoreEvent::Tick => self.on_tick(now, &mut out),
+        }
+        out
+    }
+
+    fn endpoint_for(&self, transport: TransportKind) -> Option<NbdEndpoint> {
+        match transport {
+            TransportKind::NbdUnix => self
+                .endpoints
+                .nbd_unix
+                .clone()
+                .map(|path| NbdEndpoint::Unix { path }),
+            TransportKind::NbdTcp => self
+                .endpoints
+                .nbd_tcp
+                .clone()
+                .map(|(host, port)| NbdEndpoint::Tcp { host, port }),
+            // WinDrive is lease-only (SPEC windows-swap-driver DT-7); no NBD endpoint.
+            TransportKind::WinDrive | TransportKind::DccAgent => None,
+        }
+    }
+
+    /// Sends `msg` to the tenant's live session (or logs if absent — DT-20).
+    fn to_tenant(&self, tenant: TenantId, msg: Msg, out: &mut Vec<Outbound>) {
+        match self.tenants.get(&tenant).and_then(|t| t.sid) {
+            Some(sid) => out.push(Outbound::ToSession(sid, msg)),
+            None => out.push(Outbound::Log(format!(
+                "[ramsharedd] WARN tenant {tenant} ausente; msg descartada"
+            ))),
+        }
+    }
+
+    fn slices_of(&self, tenant: TenantId) -> Vec<SliceId> {
+        self.slice_map
+            .slices()
+            .iter()
+            .filter(|s| s.tenant == Some(tenant))
+            .map(|s| s.id)
+            .collect()
+    }
+
+    fn on_msg(&mut self, sid: usize, msg: Msg, out: &mut Vec<Outbound>) {
+        match msg {
+            ramshared_broker::protocol::Msg::Register {
+                proto,
+                tenant,
+                transport,
+            } => self.on_register(sid, proto, tenant, transport, out),
+            ramshared_broker::protocol::Msg::Psi { sample, swaps, mem } => self.on_psi(sid, sample, swaps, mem, out),
+            ramshared_broker::protocol::Msg::SwapOnDone { slice, ok, detail } => {
+                if ok {
+                    out.push(Outbound::Log(format!(
+                        "[ramsharedd] swapon ok slice=s{slice}"
+                    )));
+                } else {
+                    // mount failed: nothing was written → returns the slice directly (without zero).
+                    let _ = self
+                        .slice_map
+                        .drain(slice)
+                        .and_then(|()| self.slice_map.release(slice));
+                    self.pending_dest.remove(&slice);
+                    out.push(Outbound::Log(format!(
+                        "[ramsharedd] swapon FALHOU slice=s{slice} ({detail}); slice liberada"
+                    )));
+                }
+            }
+            ramshared_broker::protocol::Msg::SwapOffDone { slice, ok, detail } => {
+                if ok {
+                    // hygiene (DT-17): zeroes before release. ZeroDone → release. The tenant already did
+                    // swapoff → safe to zero; enters pending_zero for the tick retry (R4).
+                    if let Some(s) = self.slice_map.get(slice) {
+                        let (base, len) = (s.offset, s.len);
+                        self.pending_zero.entry(slice).or_insert(0);
+                        out.push(Outbound::ZeroSlice { slice, base, len });
+                    }
+                } else {
+                    out.push(Outbound::Log(format!(
+                        "[ramsharedd] swapoff FALHOU slice=s{slice} ({detail}); fica Draining"
+                    )));
+                }
+            }
+            ramshared_broker::protocol::Msg::Status => out.push(Outbound::ToSession(sid, self.status_reply())),
+            ramshared_broker::protocol::Msg::LeaseRequest { bytes } => self.on_lease_request(sid, bytes, out),
+            ramshared_broker::protocol::Msg::LeaseRelease { lease } => self.on_lease_release(sid, lease, out),
+            other => {
+                // Broker→agent messages never arrive here; unexpected shape → close.
+                out.push(Outbound::ToSession(
+                    sid,
+                    ramshared_broker::protocol::Msg::Error {
+                        reason: format!("mensagem inesperada do agente: {other:?}"),
+                    },
+                ));
+                out.push(Outbound::CloseSession(sid));
+            }
+        }
+    }
+
+    fn on_register(
+        &mut self,
+        sid: usize,
+        proto: u32,
+        name: String,
+        transport: TransportKind,
+        out: &mut Vec<Outbound>,
+    ) {
+        if proto != PROTO_VERSION {
+            out.push(Outbound::ToSession(
+                sid,
+                ramshared_broker::protocol::Msg::Error {
+                    reason: format!("proto {proto} != {PROTO_VERSION}"),
+                },
+            ));
+            out.push(Outbound::CloseSession(sid));
+            return;
+        }
+        let id = *self
+            .name_to_id
+            .entry(name.clone())
+            .or_insert(self.next_tenant);
+        if id == self.next_tenant {
+            self.next_tenant += 1;
+        }
+        // DT-22: tenant already has a different live session → duplicate.
+        if let Some(t) = self.tenants.get(&id)
+            && t.present
+            && t.sid.is_some()
+            && t.sid != Some(sid)
+        {
+            out.push(Outbound::ToSession(
+                sid,
+                ramshared_broker::protocol::Msg::Error {
+                    reason: "tenant_duplicado".into(),
+                },
+            ));
+            out.push(Outbound::CloseSession(sid));
+            return;
+        }
+        self.tenants.insert(
+            id,
+            TenantState {
+                name: name.clone(),
+                transport,
+                present: true,
+                sid: Some(sid),
+                psi: self
+                    .tenants
+                    .get(&id)
+                    .map_or(PsiSample::default(), |t| t.psi),
+                reconciled: false,
+                mem: None,
+                occupied_bytes: 0,
+            },
+        );
+        self.sessions.insert(sid, id);
+        out.push(Outbound::Log(format!(
+            "[ramsharedd] tenant registrado name={name} id={id} transport={transport:?}"
+        )));
+        out.push(Outbound::ToSession(sid, ramshared_broker::protocol::Msg::Registered { tenant_id: id }));
+    }
+
+    fn on_psi(
+        &mut self,
+        sid: usize,
+        sample: PsiSample,
+        swaps: Vec<SwapEntry>,
+        mem: Option<TenantMem>,
+        out: &mut Vec<Outbound>,
+    ) {
+        let Some(&id) = self.sessions.get(&sid) else {
+            out.push(Outbound::ToSession(
+                sid,
+                ramshared_broker::protocol::Msg::Error {
+                    reason: "psi antes de register".into(),
+                },
+            ));
+            out.push(Outbound::CloseSession(sid));
+            return;
+        };
+        if let Some(t) = self.tenants.get_mut(&id) {
+            t.psi = sample;
+            t.mem = mem;
+            // Reconciliation (DT-9/21): on the 1st Psi, re-adopts slices that the agent already has mounted.
+            if !t.reconciled {
+                t.reconciled = true;
+                for e in &swaps {
+                    if let Some(slice) = dev_to_slice(&e.dev)
+                        && self
+                            .slice_map
+                            .get(slice)
+                            .is_some_and(|s| s.state == SliceState::Free)
+                        && self.slice_map.assign(slice, id).is_ok()
+                    {
+                        out.push(Outbound::Log(format!(
+                            "[ramsharedd] reconciliado slice=s{slice} -> tenant {id} (já montada)"
+                        )));
+                    }
+                }
+            }
+        }
+        // occupied (DT-10/F-v2-2): Σ used_kb*1024 of swaps that match Active slices of this tenant.
+        let occupied: u64 = swaps
+            .iter()
+            .filter_map(|e| dev_to_slice(&e.dev).map(|sl| (sl, e.used_kb)))
+            .filter(|(sl, _)| {
+                self.slice_map
+                    .get(*sl)
+                    .is_some_and(|s| s.state == SliceState::Active && s.tenant == Some(id))
+            })
+            .map(|(_, kb)| kb.saturating_mul(1024))
+            .fold(0_u64, u64::saturating_add);
+        if let Some(t) = self.tenants.get_mut(&id) {
+            t.occupied_bytes = occupied;
+        }
+        out.push(Outbound::ToSession(sid, ramshared_broker::protocol::Msg::Ack)); // DT-18: heartbeat
+    }
+
+    fn on_lease_request(&mut self, sid: usize, bytes: u64, out: &mut Vec<Outbound>) {
+        let Some(&holder) = self.sessions.get(&sid) else {
+            out.push(Outbound::ToSession(
+                sid,
+                ramshared_broker::protocol::Msg::Error {
+                    reason: "lease antes de register".into(),
+                },
+            ));
+            out.push(Outbound::CloseSession(sid));
+            return;
+        };
+        match self.lease_book.begin_request(holder, bytes) {
+            LeaseDecision::Pending(_) => out.push(Outbound::Log(format!(
+                "[ramsharedd] lease pedido holder={holder} bytes={bytes} (grant no próximo tick)"
+            ))),
+            LeaseDecision::Denied(reason) => out.push(Outbound::ToSession(
+                sid,
+                ramshared_broker::protocol::Msg::LeaseDenied {
+                    reason: match reason {
+                        LeaseDeny::ZeroBytes => "bytes_zero",
+                        LeaseDeny::OverCapacity => "acima_da_capacidade",
+                        _ => "lease_em_andamento",
+                    }
+                    .into(),
+                },
+            )),
+        }
+    }
+
+    fn on_lease_release(&mut self, sid: usize, lease_id: u32, out: &mut Vec<Outbound>) {
+        let Some(&holder) = self.sessions.get(&sid) else {
+            out.push(Outbound::ToSession(
+                sid,
+                ramshared_broker::protocol::Msg::Error {
+                    reason: "lease release antes de register".into(),
+                },
+            ));
+            out.push(Outbound::CloseSession(sid));
+            return;
+        };
+        match self.lease_book.release(holder, lease_id) {
+            Ok(true) => {}
+            Ok(false) | Err(LeaseDeny::WrongLease) => {
+                out.push(Outbound::Log(format!(
+                    "[ramsharedd] lease {lease_id} release idempotente/ignorado holder={holder}"
+                )));
+                return;
+            }
+            Err(LeaseDeny::WrongHolder) => {
+                out.push(Outbound::ToSession(
+                    sid,
+                    ramshared_broker::protocol::Msg::Error {
+                        reason: "holder de lease incorreto".into(),
+                    },
+                ));
+                out.push(Outbound::CloseSession(sid));
+                return;
+            }
+            Err(_) => return,
+        }
+        let leased: Vec<SliceId> = self
+            .slice_map
+            .slices()
+            .iter()
+            .filter(|s| s.state == SliceState::Leased)
+            .map(|s| s.id)
+            .collect();
+        for slice in leased {
+            let _ = self.slice_map.unlease(slice); // Leased → Free (round-robin re-leases)
+        }
+        out.push(Outbound::Log(format!(
+            "[ramsharedd] lease {lease_id} liberado; slices devolvidas ao tier de swap"
+        )));
+    }
+
+    fn on_disconnect(&mut self, sid: usize, out: &mut Vec<Outbound>) {
+        if let Some(id) = self.sessions.remove(&sid) {
+            if let Some(t) = self.tenants.get_mut(&id) {
+                t.present = false;
+                t.sid = None;
+            }
+            out.push(Outbound::Log(format!(
+                "[ramsharedd] tenant {id} desconectou; slices congeladas (DT-20)"
+            )));
+            // Lease holder/requester dropped → automatic release/cancel (DT-19).
+            let disconnected = self.lease_book.disconnect(id);
+            if let Some(lease) = disconnected.released {
+                self.release_slices(lease.id, out);
+            }
+        }
+    }
+
+    fn release_slices(&mut self, lease_id: u32, out: &mut Vec<Outbound>) {
+        let leased: Vec<SliceId> = self
+            .slice_map
+            .slices()
+            .iter()
+            .filter(|s| s.state == SliceState::Leased)
+            .map(|s| s.id)
+            .collect();
+        for slice in leased {
+            let _ = self.slice_map.unlease(slice);
+        }
+        out.push(Outbound::Log(format!(
+            "[ramsharedd] lease {lease_id} liberado; slices devolvidas ao tier de swap"
+        )));
+    }
+
+    fn on_zero_done(&mut self, slice: SliceId, ok: bool, out: &mut Vec<Outbound>) {
+        if !ok {
+            // retry zeroing on the next movement event; for now log and remain Draining (R4).
+            out.push(Outbound::Log(format!(
+                "[ramsharedd] zero FALHOU slice=s{slice}; fica Draining (retry)"
+            )));
+            if let Some(s) = self.slice_map.get(slice) {
+                out.push(Outbound::ZeroSlice {
+                    slice,
+                    base: s.offset,
+                    len: s.len,
+                });
+            }
+            return;
+        }
+        self.pending_zero.remove(&slice); // confirmed zero ends the retry (R4)
+        if self.slice_map.release(slice).is_err() {
+            return; // was not Draining; ignore
+        }
+        // Movement in flight: the cleaned slice goes to the destination (SwapOn).
+        if let Some(dest) = self.pending_dest.remove(&slice)
+            && self.slice_map.assign(slice, dest).is_ok()
+        {
+            self.emit_swapon(slice, dest, out);
+        }
+    }
+
+    fn on_demote(&mut self, reason: &str, out: &mut Vec<Outbound>) {
+        self.demotes_total = self.demotes_total.saturating_add(1);
+        self.last_demote_reason = Some(reason.to_string());
+        out.push(Outbound::Log(format!(
+            "[ramsharedd] DemoteAll reason={reason}"
+        )));
+        let present: Vec<TenantId> = self
+            .tenants
+            .iter()
+            .filter(|(_, t)| t.present)
+            .map(|(id, _)| *id)
+            .collect();
+        for id in present {
+            self.to_tenant(id, ramshared_broker::protocol::Msg::DemoteAll, out);
+        }
+    }
+
+    fn emit_swapon(&self, slice: SliceId, tenant: TenantId, out: &mut Vec<Outbound>) {
+        let transport = match self.tenants.get(&tenant) {
+            Some(t) => t.transport,
+            None => return,
+        };
+        let Some(endpoint) = self.endpoint_for(transport) else {
+            out.push(Outbound::Log(format!(
+                "[ramsharedd] ERRO transporte indisponível p/ tenant {tenant}"
+            )));
+            return;
+        };
+        self.to_tenant(
+            tenant,
+            ramshared_broker::protocol::Msg::SwapOn {
+                slice,
+                export: format!("s{slice}"),
+                endpoint,
+                swap_prio: self.swap_prio,
+            },
+            out,
+        );
+    }
+
+    fn status_reply(&self) -> Msg {
+        let tenants = self
+            .tenants
+            .iter()
+            .map(|(id, t)| TenantStatus {
+                id: *id,
+                name: t.name.clone(),
+                psi: t.psi,
+                slices: self.slices_of(*id),
+                present: t.present,
+                bytes_served: self
+                    .slices_of(*id)
+                    .iter()
+                    .filter_map(|s| self.slice_io.get(*s as usize))
+                    .map(|c| c.bytes_served.load(Ordering::Relaxed))
+                    .fold(0_u64, u64::saturating_add),
+            })
+            .collect();
+        let slice_io = self
+            .slice_map
+            .slices()
+            .iter()
+            .filter_map(|s| {
+                self.slice_io.get(s.id as usize).map(|c| SliceIo {
+                    id: s.id,
+                    bytes_served: c.bytes_served.load(Ordering::Relaxed),
+                    io_count: c.io_count.load(Ordering::Relaxed),
+                })
+            })
+            .collect();
+        ramshared_broker::protocol::Msg::StatusReply {
+            tenants,
+            slices: self.slice_map.slices().to_vec(),
+            slice_io,
+            last_rebalance_secs: None,
+        }
+    }
+
+    fn on_tick(&mut self, now: Instant, out: &mut Vec<Outbound>) {
+        self.tick_arbiter(now, out);
+        self.retry_stuck_zeroing(out);
+        self.emit_telemetry(out);
+    }
+
+    fn tick_arbiter(&mut self, now: Instant, out: &mut Vec<Outbound>) {
+        // DT-20: only present tenants; only Free slices or slices with a present owner.
+        // WinDrive / lease-only transports are excluded from swap round-robin (DT-7).
+        let present: Vec<TenantView> = self
+            .tenants
+            .iter()
+            .filter(|(_, t)| {
+                t.present
+                    && !matches!(
+                        t.transport,
+                        TransportKind::WinDrive | TransportKind::DccAgent
+                    )
+            })
+            .map(|(id, t)| TenantView {
+                id: *id,
+                psi: t.psi,
+                slices: self.slices_of(*id).len() as u16,
+            })
+            .collect();
+        let visible: Vec<Slice> = self
+            .slice_map
+            .slices()
+            .iter()
+            .filter(|s| {
+                s.tenant.is_none()
+                    || s.tenant
+                        .is_some_and(|t| self.tenants.get(&t).is_some_and(|ts| ts.present))
+            })
+            .cloned()
+            .collect();
+
+        let pending_lease = self
+            .lease_book
+            .pending()
+            .map(|pending| (pending.holder, pending.requested_bytes));
+        let actions = match self.arbiter.tick(now, &present, &visible, pending_lease) {
+            Ok(acts) => acts,
+            Err(e) => {
+                out.push(Outbound::Log(format!(
+                    "[ramsharedd] arbiter tick error: {e}"
+                )));
+                Vec::new()
+            }
+        };
+        for action in actions {
+            match action {
+                Action::AssignFree { slice, to } => {
+                    if self.slice_map.assign(slice, to).is_ok() {
+                        out.push(Outbound::Log(format!(
+                            "[ramsharedd] arbiter assign slice=s{slice} to={to}"
+                        )));
+                        self.emit_swapon(slice, to, out);
+                    }
+                }
+                Action::MoveSlice { slice, from, to } | Action::RevertMove { slice, from, to } => {
+                    if self.slice_map.drain(slice).is_ok() {
+                        self.pending_dest.insert(slice, to);
+                        self.last_rebalance = Some(now);
+                        let pf = present
+                            .iter()
+                            .find(|t| t.id == from)
+                            .map_or(0.0, |t| t.psi.avg10);
+                        let pt = present
+                            .iter()
+                            .find(|t| t.id == to)
+                            .map_or(0.0, |t| t.psi.avg10);
+                        out.push(Outbound::Log(format!(
+                            "[ramsharedd] arbiter move slice=s{slice} from={from}(psi10={pf:.1}) to={to}(psi10={pt:.1})"
+                        )));
+                        self.to_tenant(from, ramshared_broker::protocol::Msg::SwapOff { slice }, out);
+                    }
+                }
+                Action::RevokeForLease { slice, from, .. } => {
+                    // Revokes for the lease: drain+SwapOff; on ZeroDone becomes Free (no pending_dest),
+                    // and the arbiter counts it for the lease on the next tick (DT-8/R2).
+                    if self.slice_map.drain(slice).is_ok() {
+                        out.push(Outbound::Log(format!(
+                            "[ramsharedd] revoke-for-lease slice=s{slice} from={from}"
+                        )));
+                        self.to_tenant(from, ramshared_broker::protocol::Msg::SwapOff { slice }, out);
+                    }
+                }
+                Action::GrantLease { holder, slices } => {
+                    let mut granted = 0u64;
+                    for s in &slices {
+                        if let Some(len) = self.slice_map.get(*s).map(|sl| sl.len)
+                            && self.slice_map.lease(*s).is_ok()
+                        {
+                            granted += len; // Free → Leased
+                        }
+                    }
+                    let Ok(lease) = self.lease_book.grant_pending(granted) else {
+                        out.push(Outbound::Log(
+                            "[ramsharedd] ERRO lease grant divergiu do LeaseBook".into(),
+                        ));
+                        continue;
+                    };
+                    out.push(Outbound::Log(format!(
+                        "[ramsharedd] lease {} concedido holder={holder} slices={slices:?}",
+                        lease.id
+                    )));
+                    self.to_tenant(
+                        holder,
+                        ramshared_broker::protocol::Msg::LeaseGranted {
+                            lease: lease.id,
+                            bytes: granted,
+                        },
+                        out,
+                    );
+                }
+            }
+        }
+    }
+
+    fn retry_stuck_zeroing(&mut self, out: &mut Vec<Outbound>) {
+        // R4: retries zeroing of stuck slices (if `try_send(ZeroExport)` failed, `ZeroDone` doesn't arrive).
+        // Grace period of 1 tick for zero in flight; above that, re-emits; ERROR after N
+        // ticks without confirmation. Only touches slices in `pending_zero` (already swapped-off; safe to re-zero).
+        let stuck: Vec<SliceId> = self.pending_zero.keys().copied().collect();
+        for slice in stuck {
+            let count = match self.pending_zero.get_mut(&slice) {
+                Some(n) => {
+                    *n += 1;
+                    *n
+                }
+                None => continue,
+            };
+            if count == ZERO_RETRY_ERROR {
+                out.push(Outbound::Log(format!(
+                    "[ramsharedd] ERRO: zero de s{slice} sem confirmar em {ZERO_RETRY_ERROR} ticks (slice presa em Draining, R4); re-tentando"
+                )));
+            }
+            if count > ZERO_RETRY_GRACE
+                && let Some(s) = self.slice_map.get(slice)
+            {
+                let (base, len) = (s.offset, s.len);
+                out.push(Outbound::ZeroSlice { slice, base, len });
+            }
+        }
+    }
+
+    /// Reconciliation per tick (RF-4/RF-5): occupancy invariant + hysteresis (DT-12) → emits
+    /// `Outbound::Telemetry`. Observer: does not touch the arbiter/SliceMap.
+    fn emit_telemetry(&mut self, out: &mut Vec<Outbound>) {
+        let alloc_active: u64 = self
+            .slice_map
+            .slices()
+            .iter()
+            .filter(|s| matches!(s.state, SliceState::Active | SliceState::Draining))
+            .map(|s| s.len)
+            .fold(0_u64, u64::saturating_add);
+        let occupied: u64 = self
+            .tenants
+            .values()
+            .map(|t| t.occupied_bytes)
+            .fold(0_u64, u64::saturating_add);
+        // Σ diskstats (cumulative in bytes) of tenants reporting `mem` (RF-2); `None` if nobody
+        // reports. The consumer derives the rate by the difference between samples (`t` field of the line).
+        let page_io: Option<u64> = self.tenants.values().any(|t| t.mem.is_some()).then(|| {
+            self.tenants
+                .values()
+                .filter_map(|t| t.mem.as_ref())
+                .map(|m| m.diskstats_io)
+                .fold(0_u64, u64::saturating_add)
+        });
+        let free = self.vram.free.load(Ordering::Relaxed);
+        let total = self.vram.total.load(Ordering::Relaxed);
+        let has_vram = total > 0; // F-v2-6: RAM sentinel (no GPU) → vram_* = None
+        let vram_alloc_daemon = alloc_active.saturating_add(CANARY_BYTES as u64);
+        let vram_total_used = has_vram.then(|| total.saturating_sub(free));
+        let vram_others = vram_total_used.map(|u| vram_outros(u, vram_alloc_daemon));
+        let demotes_delta = self
+            .demotes_total
+            .saturating_sub(self.demotes_at_last_sample);
+        self.demotes_at_last_sample = self.demotes_total;
+        let stuck_draining = self.pending_zero.values().any(|n| *n >= ZERO_RETRY_ERROR);
+        let any_present = self.tenants.values().any(|t| t.present);
+        let inp = ReconcileInput {
+            alloc_active_bytes: alloc_active,
+            occupied_swap_bytes: occupied,
+            stuck_draining,
+            demotes_delta,
+            any_source_missing: !has_vram || !any_present,
+        };
+        let (delta, candidate) = reconcile(&inp, self.tol_frac);
+        // Hysteresis (DT-12) for SUSTAINED flags (Unaccounted/StuckSlice/Partial): confirms after
+        // `recon_streak` consecutive identical ticks. `Eviction` is a canary EVENT (DT-6;
+        // `demotes_delta` is per-tick, lasts 1 tick) → immediate confirmation; otherwise hysteresis would swallow
+        // a transient eviction (1-2 DEMOTEs) and the eviction signal would never appear (bug C1).
+        if candidate == self.recon_flag {
+            self.recon_count = self.recon_count.saturating_add(1);
+        } else {
+            self.recon_flag = candidate;
+            self.recon_count = 1;
+        }
+        let confirmed = match candidate {
+            ReconcileFlag::Eviction => ReconcileFlag::Eviction, // evento: sem histerese
+            ReconcileFlag::None => ReconcileFlag::None,
+            sustained if self.recon_count >= self.recon_streak => sustained,
+            _ => ReconcileFlag::None,
+        };
+        if confirmed != ReconcileFlag::None {
+            out.push(Outbound::Log(format!(
+                "[ramsharedd] reconcile flag={confirmed:?} delta={delta:.3} reason={:?}",
+                self.last_demote_reason
+            )));
+        }
+        out.push(Outbound::Telemetry(TelemetryCore {
+            tenant: None,
+            slice: None,
+            swap_used: occupied,
+            alloc_active,
+            page_io_s: page_io,
+            vram_alloc_daemon,
+            vram_total_used,
+            vram_outros: vram_others,
+            canario_demotes: self.demotes_total,
+            demote_reason: self.last_demote_reason.clone(),
+            reconcile_delta: delta,
+            flag: confirmed,
+        }));
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    #![allow(clippy::unwrap_used, clippy::expect_used)]
+    use ramshared_broker::arbiter::ArbiterConfig;
+    use ramshared_broker::slices::SliceMap;
+    use super::super::BrokerCoreConfig;
+    use std::sync::Arc;
+    use crate::telemetry::SliceIoCounters;
+    use crate::telemetry::VramGauge;
+    use super::super::TelemetrySink;
+    use std::path::PathBuf;
+    use super::*;
+    use ramshared_broker::model::SliceState;
+
+    fn core(k: u16) -> BrokerCore {
+        core_streak(k, 1)
+    }
+
+    /// Like `core`, but with configurable `recon_streak` (tests the actual production hysteresis).
+    fn core_streak(k: u16, recon_streak: u32) -> BrokerCore {
+        let cfg = ArbiterConfig {
+            streak: 1, // move already on the 1st tick above delta (tests)
+            ..ArbiterConfig::default()
+        };
+        BrokerCore::new(
+            SliceMap::new(k, 64 * 1024 * 1024, u64::from(k) * 64 * 1024 * 1024).unwrap(),
+            BrokerCoreConfig {
+                arbiter_cfg: cfg,
+                endpoints: EndpointCfg {
+                    nbd_unix: Some("/run/x.sock".into()),
+                    nbd_tcp: None,
+                },
+                swap_prio: None,
+                slice_io: Arc::new((0..k).map(|_| SliceIoCounters::default()).collect()),
+                vram: Arc::new(VramGauge::default()),
+                tol_frac: 0.10,
+                recon_streak,
+            },
+        )
+    }
+
+    fn reg(c: &mut BrokerCore, sid: usize, name: &str) -> Vec<Outbound> {
+        c.handle(
+            CoreEvent::Msg(
+                sid,
+                ramshared_broker::protocol::Msg::Register {
+                    proto: PROTO_VERSION,
+                    tenant: name.into(),
+                    transport: TransportKind::NbdUnix,
+                },
+            ),
+            Instant::now(),
+        )
+    }
+
+    fn psi(c: &mut BrokerCore, sid: usize, avg10: f32) -> Vec<Outbound> {
+        c.handle(
+            CoreEvent::Msg(
+                sid,
+                ramshared_broker::protocol::Msg::Psi {
+                    sample: PsiSample {
+                        avg10,
+                        avg60: avg10,
+                        stall_us: 0,
+                    },
+                    swaps: vec![],
+                    mem: None,
+                },
+            ),
+            Instant::now(),
+        )
+    }
+
+    #[test]
+    fn register_assigns_stable_id_and_acks_psi() {
+        let mut c = core(2);
+        let o = reg(&mut c, 10, "wsl2");
+        assert!(o.contains(&Outbound::ToSession(10, ramshared_broker::protocol::Msg::Registered { tenant_id: 1 })));
+        let o = psi(&mut c, 10, 0.0);
+        assert!(o.contains(&Outbound::ToSession(10, ramshared_broker::protocol::Msg::Ack)));
+    }
+
+    #[test]
+    fn duplicate_register_is_rejected() {
+        let mut c = core(2);
+        reg(&mut c, 10, "wsl2");
+        let o = reg(&mut c, 11, "wsl2"); // same name, another live connection
+        assert!(o.iter().any(|x| matches!(x, Outbound::CloseSession(11))));
+        assert!(o.iter().any(
+            |x| matches!(x, Outbound::ToSession(11, ramshared_broker::protocol::Msg::Error { reason }) if reason == "tenant_duplicado")
+        ));
+    }
+
+    #[test]
+    fn proto_mismatch_rejected() {
+        let mut c = core(1);
+        let o = c.handle(
+            CoreEvent::Msg(
+                10,
+                ramshared_broker::protocol::Msg::Register {
+                    proto: 999,
+                    tenant: "x".into(),
+                    transport: TransportKind::NbdUnix,
+                },
+            ),
+            Instant::now(),
+        );
+        assert!(o.iter().any(|x| matches!(x, Outbound::CloseSession(10))));
+    }
+
+    #[test]
+    fn psi_before_register_closes() {
+        let mut c = core(1);
+        let o = psi(&mut c, 7, 0.0);
+        assert!(o.iter().any(|x| matches!(x, Outbound::CloseSession(7))));
+    }
+
+    #[test]
+    fn tick_round_robin_assigns_free_slices() {
+        let mut c = core(2);
+        reg(&mut c, 10, "a");
+        reg(&mut c, 20, "b");
+        psi(&mut c, 10, 0.0);
+        psi(&mut c, 20, 0.0);
+        let o = c.handle(CoreEvent::Tick, Instant::now());
+        // 2 Free slices → SwapOn for each tenant (round-robin: s0→a(1), s1→b(2))
+        let swapons: Vec<_> = o
+            .iter()
+            .filter(|x| matches!(x, Outbound::ToSession(_, ramshared_broker::protocol::Msg::SwapOn { .. })))
+            .collect();
+        assert_eq!(swapons.len(), 2);
+        assert_eq!(c.slice_map.get(0).unwrap().state, SliceState::Active);
+        assert_eq!(c.slice_map.get(1).unwrap().state, SliceState::Active);
+    }
+
+    #[test]
+    fn swapoff_done_triggers_zero_then_release() {
+        let mut c = core(1);
+        reg(&mut c, 10, "a");
+        psi(&mut c, 10, 0.0);
+        c.handle(CoreEvent::Tick, Instant::now()); // assign s0 → a (Active)
+        c.slice_map.drain(0).unwrap(); // simulates start of move (Active→Draining)
+        let o = c.handle(
+            CoreEvent::Msg(
+                10,
+                ramshared_broker::protocol::Msg::SwapOffDone {
+                    slice: 0,
+                    ok: true,
+                    detail: String::new(),
+                },
+            ),
+            Instant::now(),
+        );
+        assert!(o.iter().any(|x| matches!(
+            x,
+            Outbound::ZeroSlice { slice: 0, base: 0, len } if *len == 64 * 1024 * 1024
+        )));
+        // ZeroDone → release → Free
+        c.handle(CoreEvent::ZeroDone(0, true), Instant::now());
+        assert_eq!(c.slice_map.get(0).unwrap().state, SliceState::Free);
+    }
+
+    #[test]
+    fn stuck_draining_zero_is_retried_on_tick() {
+        // R4: if zeroing is not confirmed (channel full → no ZeroDone), the tick re-emits ZeroExport
+        // after grace and escalates to ERROR; once confirmed, stops retrying.
+        let mut c = core(1);
+        reg(&mut c, 10, "a");
+        psi(&mut c, 10, 0.0);
+        c.handle(CoreEvent::Tick, Instant::now()); // assign s0 (Active)
+        c.slice_map.drain(0).unwrap(); // Draining
+        c.handle(
+            CoreEvent::Msg(
+                10,
+                ramshared_broker::protocol::Msg::SwapOffDone {
+                    slice: 0,
+                    ok: true,
+                    detail: String::new(),
+                },
+            ),
+            Instant::now(),
+        );
+        // WITHOUT ZeroDone (simulates full channel). Tick 1 = grace → does not re-emit.
+        let t1 = c.handle(CoreEvent::Tick, Instant::now());
+        assert!(
+            !t1.iter()
+                .any(|x| matches!(x, Outbound::ZeroSlice { slice: 0, .. })),
+            "carência: sem retry no 1º tick"
+        );
+        // Tick 2 → re-emits the zero.
+        let t2 = c.handle(CoreEvent::Tick, Instant::now());
+        assert!(
+            t2.iter()
+                .any(|x| matches!(x, Outbound::ZeroSlice { slice: 0, .. })),
+            "retry do zero no 2º tick"
+        );
+        // More ticks → escalates to ERROR (R4).
+        let mut saw_error = false;
+        for _ in 0..ZERO_RETRY_ERROR {
+            let o = c.handle(CoreEvent::Tick, Instant::now());
+            if o.iter()
+                .any(|x| matches!(x, Outbound::Log(s) if s.contains("R4")))
+            {
+                saw_error = true;
+            }
+        }
+        assert!(
+            saw_error,
+            "must escalate to ERROR (R4) after N unconfirmed ticks"
+        );
+        // Zero confirms → Free + stops retrying.
+        c.handle(CoreEvent::ZeroDone(0, true), Instant::now());
+        assert_eq!(c.slice_map.get(0).unwrap().state, SliceState::Free);
+        let after = c.handle(CoreEvent::Tick, Instant::now());
+        assert!(
+            !after
+                .iter()
+                .any(|x| matches!(x, Outbound::ZeroSlice { slice: 0, .. })),
+            "after confirmation: no more retries"
+        );
+    }
+
+    #[test]
+    fn move_drains_donor_then_swapon_dest_after_zero() {
+        let mut c = core(2);
+        reg(&mut c, 10, "donor");
+        reg(&mut c, 20, "recv");
+        psi(&mut c, 10, 0.0);
+        psi(&mut c, 20, 0.0);
+        c.handle(CoreEvent::Tick, Instant::now()); // assign s0→donor(1), s1→recv(2)
+        // imbalance: recv (2) highly pressured, donor (1) idle, no Free slices → move
+        psi(&mut c, 10, 0.0);
+        psi(&mut c, 20, 50.0);
+        let o = c.handle(CoreEvent::Tick, Instant::now());
+        // move de uma slice do donor(1) p/ recv(2): SwapOff p/ a sessão do donor (10)
+        assert!(
+            o.iter()
+                .any(|x| matches!(x, Outbound::ToSession(10, ramshared_broker::protocol::Msg::SwapOff { .. })))
+        );
+        let moved = c
+            .slice_map
+            .slices()
+            .iter()
+            .find(|s| s.state == SliceState::Draining)
+            .map(|s| s.id)
+            .expect("uma slice drenando");
+        // SwapOffDone → zero → ZeroDone → SwapOn for recv (session 20)
+        c.handle(
+            CoreEvent::Msg(
+                10,
+                ramshared_broker::protocol::Msg::SwapOffDone {
+                    slice: moved,
+                    ok: true,
+                    detail: String::new(),
+                },
+            ),
+            Instant::now(),
+        );
+        let o = c.handle(CoreEvent::ZeroDone(moved, true), Instant::now());
+        assert!(o.iter().any(|x| matches!(
+            x,
+            Outbound::ToSession(20, ramshared_broker::protocol::Msg::SwapOn { slice, .. }) if *slice == moved
+        )));
+        assert_eq!(c.slice_map.get(moved).unwrap().tenant, Some(2));
+    }
+
+    #[test]
+    fn disconnect_marks_absent_and_freezes() {
+        let mut c = core(1);
+        reg(&mut c, 10, "a");
+        psi(&mut c, 10, 0.0);
+        c.handle(CoreEvent::Tick, Instant::now()); // s0 → a (Active)
+        c.handle(CoreEvent::Disconnected(10), Instant::now());
+        // slice remains Active (frozen), owner absent
+        assert_eq!(c.slice_map.get(0).unwrap().state, SliceState::Active);
+        // tick does not touch the absent one's slice (does not become Free nor reassigned)
+        let o = c.handle(CoreEvent::Tick, Instant::now());
+        assert!(
+            !o.iter()
+                .any(|x| matches!(x, Outbound::ToSession(_, ramshared_broker::protocol::Msg::SwapOn { .. })))
+        );
+        assert_eq!(c.slice_map.get(0).unwrap().state, SliceState::Active);
+    }
+
+    #[test]
+    fn demote_broadcasts_to_present() {
+        let mut c = core(1);
+        reg(&mut c, 10, "a");
+        let o = c.handle(CoreEvent::Demote("latency".into()), Instant::now());
+        assert!(o.contains(&Outbound::ToSession(10, ramshared_broker::protocol::Msg::DemoteAll)));
+    }
+
+    #[test]
+    fn status_reply_lists_tenants_and_slices() {
+        let mut c = core(2);
+        reg(&mut c, 10, "a");
+        let o = c.handle(CoreEvent::Msg(10, ramshared_broker::protocol::Msg::Status), Instant::now());
+        assert!(o.iter().any(|x| matches!(
+            x,
+            Outbound::ToSession(10, ramshared_broker::protocol::Msg::StatusReply { tenants, slices, .. })
+                if tenants.len() == 1 && slices.len() == 2
+        )));
+    }
+
+    #[test]
+    fn status_reply_includes_slice_io() {
+        // RF-1: StatusReply exposes counters per slice (read from the shared Arc).
+        let c = core(2);
+        c.slice_io[0].bytes_served.store(4096, Ordering::Relaxed);
+        c.slice_io[0].io_count.store(1, Ordering::Relaxed);
+        match c.status_reply() {
+            ramshared_broker::protocol::Msg::StatusReply { slice_io, .. } => {
+                assert_eq!(slice_io.len(), 2);
+                let s0 = slice_io.iter().find(|s| s.id == 0).expect("slice 0");
+                assert_eq!(s0.bytes_served, 4096);
+                assert_eq!(s0.io_count, 1);
+            }
+            other => panic!("esperava StatusReply, veio {other:?}"),
+        }
+    }
+
+    #[test]
+    fn on_tick_emits_telemetry() {
+        // RF-5: every tick emits a telemetry sample.
+        let mut c = core(1);
+        let o = c.handle(CoreEvent::Tick, Instant::now());
+        assert!(o.iter().any(|x| matches!(x, Outbound::Telemetry(_))));
+    }
+
+    #[test]
+    fn eviction_flag_after_demote() {
+        // RF-4/DT-6: a canary DEMOTE → Eviction flag on the next tick (streak=1 in helper).
+        let mut c = core(1);
+        reg(&mut c, 10, "a");
+        c.vram.total.store(1 << 30, Ordering::Relaxed); // has_vram (otherwise Partial)
+        c.vram.free.store(1 << 29, Ordering::Relaxed);
+        c.handle(CoreEvent::Demote("latency".into()), Instant::now());
+        let o = c.handle(CoreEvent::Tick, Instant::now());
+        let flag = o
+            .iter()
+            .find_map(|x| match x {
+                Outbound::Telemetry(t) => Some(t.flag),
+                _ => None,
+            })
+            .expect("telemetry on tick");
+        assert_eq!(flag, ReconcileFlag::Eviction);
+    }
+
+    #[test]
+    fn unaccounted_when_occupied_exceeds_alloc() {
+        // RF-4: occupied > borrowed + tol → Unaccounted (no demote, with VRAM).
+        let mut c = core(1); // 1 slice of 64 MiB
+        reg(&mut c, 10, "a");
+        let id = *c.tenants.keys().next().expect("tenant");
+        c.slice_map.assign(0, id).expect("assign slice 0"); // Free→Active (64 MiB borrowed)
+        c.tenants.get_mut(&id).expect("tenant").occupied_bytes = 200 * 1024 * 1024; // 200 MiB > 64
+        c.vram.total.store(1 << 30, Ordering::Relaxed); // has_vram
+        let o = c.handle(CoreEvent::Tick, Instant::now());
+        let flag = o
+            .iter()
+            .find_map(|x| match x {
+                Outbound::Telemetry(t) => Some(t.flag),
+                _ => None,
+            })
+            .expect("telemetry on tick");
+        assert_eq!(flag, ReconcileFlag::Unaccounted);
+    }
+
+    fn tick_flag(c: &mut BrokerCore) -> ReconcileFlag {
+        c.handle(CoreEvent::Tick, Instant::now())
+            .iter()
+            .find_map(|x| match x {
+                Outbound::Telemetry(t) => Some(t.flag),
+                _ => None,
+            })
+            .expect("telemetry on tick")
+    }
+
+    #[test]
+    fn eviction_confirmed_immediately_despite_streak() {
+        // C1: Eviction is an EVENT (canary) → confirms in 1 tick even with recon_streak=3 (production).
+        // Without the fix, the hysteresis would swallow transient eviction.
+        let mut c = core_streak(1, 3);
+        reg(&mut c, 10, "a");
+        c.vram.total.store(1 << 30, Ordering::Relaxed); // has_vram (otherwise Partial)
+        c.vram.free.store(1 << 29, Ordering::Relaxed);
+        c.handle(CoreEvent::Demote("latency".into()), Instant::now());
+        assert_eq!(tick_flag(&mut c), ReconcileFlag::Eviction);
+    }
+
+    #[test]
+    fn unaccounted_respects_streak() {
+        // RF-4/DT-12: SUSTAINED flag (Unaccounted) only confirms after `recon_streak` ticks.
+        let mut c = core_streak(1, 2);
+        reg(&mut c, 10, "a");
+        let id = *c.tenants.keys().next().expect("tenant");
+        c.slice_map.assign(0, id).expect("assign");
+        c.tenants.get_mut(&id).expect("tenant").occupied_bytes = 200 * 1024 * 1024;
+        c.vram.total.store(1 << 30, Ordering::Relaxed);
+        assert_eq!(
+            tick_flag(&mut c),
+            ReconcileFlag::None,
+            "first tick: streak not reached yet"
+        );
+        assert_eq!(
+            tick_flag(&mut c),
+            ReconcileFlag::Unaccounted,
+            "second tick: confirms"
+        );
+    }
+
+    #[test]
+    fn telemetry_sink_writes_jsonl_line() {
+        // RF-5/DT-8: the sink writes 1 JSON object per line (write-in-file, in-process).
+        let path =
+            std::env::temp_dir().join(format!("ramshared-telem-{}.jsonl", std::process::id()));
+        let _ = std::fs::remove_file(&path);
+        {
+            let mut sink = TelemetrySink::open(path.clone()).expect("abre sink");
+            let tc = TelemetryCore {
+                tenant: None,
+                slice: None,
+                swap_used: 7,
+                alloc_active: 8,
+                page_io_s: None,
+                vram_alloc_daemon: 9,
+                vram_total_used: None,
+                vram_outros: None,
+                canario_demotes: 0,
+                demote_reason: None,
+                reconcile_delta: 0.0,
+                flag: ReconcileFlag::None,
+            };
+            sink.emit(&tc).expect("first telemetry row");
+            sink.emit(&tc).expect("second telemetry row");
+        }
+        let content = std::fs::read_to_string(&path).expect("reads the jsonl");
+        let lines: Vec<&str> = content.lines().collect();
+        assert_eq!(lines.len(), 2, "one line per emit");
+        let v: serde_json::Value = serde_json::from_str(lines[0]).expect("valid JSON");
+        assert_eq!(v["swap_used"], 7);
+        assert!(v.get("t").is_some(), "timestamp is present");
+        let _ = std::fs::remove_file(&path);
+    }
+
+    #[cfg(target_os = "linux")]
+    #[test]
+    fn telemetry_sink_surfaces_write_failure() {
+        let mut sink = TelemetrySink::open(PathBuf::from("/dev/full")).expect("open /dev/full");
+        let tc = TelemetryCore {
+            tenant: None,
+            slice: None,
+            swap_used: 0,
+            alloc_active: 0,
+            page_io_s: None,
+            vram_alloc_daemon: 0,
+            vram_total_used: None,
+            vram_outros: None,
+            canario_demotes: 0,
+            demote_reason: None,
+            reconcile_delta: 0.0,
+            flag: ReconcileFlag::Partial,
+        };
+        assert!(sink.emit(&tc).is_err());
+    }
+
+    #[test]
+    fn dev_to_slice_requires_exact_nbd_identity() {
+        assert_eq!(dev_to_slice("/dev/nbd5"), Some(5));
+        assert_eq!(dev_to_slice("/dev/nbd0"), Some(0));
+        assert_eq!(dev_to_slice("/nbd7"), Some(7));
+        assert_eq!(dev_to_slice("nbd9"), Some(9));
+        assert_eq!(dev_to_slice("/dev/sda"), None);
+        assert_eq!(dev_to_slice("/dev/sda5"), None);
+        assert_eq!(dev_to_slice("/tmp/nbd5"), None);
+        assert_eq!(dev_to_slice("/dev/nbd5-backup"), None);
+        assert_eq!(dev_to_slice("/dev/nbd5\\040(deleted)"), None);
+    }
+
+    const SLICE: u64 = 64 * 1024 * 1024;
+
+    fn lease_req(c: &mut BrokerCore, sid: usize, bytes: u64) -> Vec<Outbound> {
+        c.handle(
+            CoreEvent::Msg(sid, ramshared_broker::protocol::Msg::LeaseRequest { bytes }),
+            Instant::now(),
+        )
+    }
+
+    fn n_leased(c: &BrokerCore) -> usize {
+        c.slice_map
+            .slices()
+            .iter()
+            .filter(|s| s.state == SliceState::Leased)
+            .count()
+    }
+
+    #[test]
+    fn lease_granted_from_free_slices() {
+        let mut c = core(2);
+        reg(&mut c, 10, "dcc");
+        psi(&mut c, 10, 0.0);
+        let o = lease_req(&mut c, 10, SLICE); // 1 slice
+        assert!(
+            !o.iter()
+                .any(|x| matches!(x, Outbound::ToSession(_, ramshared_broker::protocol::Msg::LeaseDenied { .. })))
+        );
+        let o = c.handle(CoreEvent::Tick, Instant::now());
+        assert!(
+            o.iter()
+                .any(|x| matches!(x, Outbound::ToSession(10, ramshared_broker::protocol::Msg::LeaseGranted { .. })))
+        );
+        assert_eq!(n_leased(&c), 1);
+    }
+
+    #[test]
+    fn lease_denied_when_in_progress() {
+        let mut c = core(2);
+        reg(&mut c, 10, "dcc");
+        psi(&mut c, 10, 0.0);
+        lease_req(&mut c, 10, SLICE);
+        let o = lease_req(&mut c, 10, SLICE);
+        assert!(o.iter().any(
+            |x| matches!(x, Outbound::ToSession(10, ramshared_broker::protocol::Msg::LeaseDenied { reason }) if reason == "lease_em_andamento")
+        ));
+    }
+
+    #[test]
+    fn lease_denied_over_capacity() {
+        let mut c = core(2); // total = 128 MiB
+        reg(&mut c, 10, "dcc");
+        psi(&mut c, 10, 0.0);
+        let o = lease_req(&mut c, 10, 200 * 1024 * 1024);
+        assert!(o.iter().any(
+            |x| matches!(x, Outbound::ToSession(10, ramshared_broker::protocol::Msg::LeaseDenied { reason }) if reason == "acima_da_capacidade")
+        ));
+    }
+
+    #[test]
+    fn lease_release_returns_slices() {
+        let mut c = core(2);
+        reg(&mut c, 10, "dcc");
+        psi(&mut c, 10, 0.0);
+        lease_req(&mut c, 10, SLICE);
+        c.handle(CoreEvent::Tick, Instant::now()); // GrantLease (lease id 1)
+        assert_eq!(n_leased(&c), 1);
+        c.handle(
+            CoreEvent::Msg(10, ramshared_broker::protocol::Msg::LeaseRelease { lease: 1 }),
+            Instant::now(),
+        );
+        assert_eq!(n_leased(&c), 0); // devolvida ao tier de swap
+    }
+
+    #[test]
+    fn foreign_tenant_cannot_release_lease() {
+        let mut c = core(2);
+        reg(&mut c, 10, "holder");
+        reg(&mut c, 20, "foreign");
+        lease_req(&mut c, 10, SLICE);
+        c.handle(CoreEvent::Tick, Instant::now());
+
+        let out = c.handle(
+            CoreEvent::Msg(20, ramshared_broker::protocol::Msg::LeaseRelease { lease: 1 }),
+            Instant::now(),
+        );
+
+        assert_eq!(n_leased(&c), 1);
+        assert!(out.contains(&Outbound::CloseSession(20)));
+        assert!(out.iter().any(
+            |effect| matches!(effect, Outbound::ToSession(20, ramshared_broker::protocol::Msg::Error { reason }) if reason.contains("holder"))
+        ));
+    }
+
+    #[test]
+    fn shared_lease_book_preserves_linux_wire_effects() {
+        let mut c = core(2);
+        reg(&mut c, 10, "holder");
+
+        lease_req(&mut c, 10, SLICE);
+        let grant = c.handle(CoreEvent::Tick, Instant::now());
+        assert!(grant.contains(&Outbound::ToSession(
+            10,
+            ramshared_broker::protocol::Msg::LeaseGranted {
+                lease: 1,
+                bytes: SLICE,
+            },
+        )));
+
+        c.handle(
+            CoreEvent::Msg(10, ramshared_broker::protocol::Msg::LeaseRelease { lease: 1 }),
+            Instant::now(),
+        );
+        lease_req(&mut c, 10, SLICE);
+        let second = c.handle(CoreEvent::Tick, Instant::now());
+        assert!(second.contains(&Outbound::ToSession(
+            10,
+            ramshared_broker::protocol::Msg::LeaseGranted {
+                lease: 2,
+                bytes: SLICE,
+            },
+        )));
+    }
+
+    #[test]
+    fn lease_revokes_active_then_grants_after_zero() {
+        // DT-8/R2: lease drena uma slice Active (de outro tenant), zera, e concede.
+        let mut c = core(1);
+        reg(&mut c, 10, "swap"); // id 1
+        reg(&mut c, 20, "dcc"); // id 2
+        psi(&mut c, 10, 0.0);
+        psi(&mut c, 20, 0.0);
+        c.handle(CoreEvent::Tick, Instant::now()); // s0 → swap(1) Active (round-robin)
+        assert_eq!(c.slice_map.get(0).unwrap().tenant, Some(1));
+        // dcc(20) requests a lease that needs the slice → revokes it from swap(1)
+        lease_req(&mut c, 20, SLICE);
+        let o = c.handle(CoreEvent::Tick, Instant::now());
+        assert!(
+            o.iter()
+                .any(|x| matches!(x, Outbound::ToSession(10, ramshared_broker::protocol::Msg::SwapOff { slice: 0 })))
+        );
+        assert!(
+            !o.iter()
+                .any(|x| matches!(x, Outbound::ToSession(_, ramshared_broker::protocol::Msg::LeaseGranted { .. })))
+        ); // not yet
+        // swapoff confirms → zero → release → Free
+        c.handle(
+            CoreEvent::Msg(
+                10,
+                ramshared_broker::protocol::Msg::SwapOffDone {
+                    slice: 0,
+                    ok: true,
+                    detail: String::new(),
+                },
+            ),
+            Instant::now(),
+        );
+        c.handle(CoreEvent::ZeroDone(0, true), Instant::now());
+        assert_eq!(c.slice_map.get(0).unwrap().state, SliceState::Free);
+        // next tick: enough Free capacity now exists → GrantLease for dcc(20)
+        let o = c.handle(CoreEvent::Tick, Instant::now());
+        assert!(
+            o.iter()
+                .any(|x| matches!(x, Outbound::ToSession(20, ramshared_broker::protocol::Msg::LeaseGranted { .. })))
+        );
+        assert_eq!(c.slice_map.get(0).unwrap().state, SliceState::Leased);
+    }
+
+    #[test]
+    fn lease_released_when_holder_disconnects() {
+        let mut c = core(2);
+        reg(&mut c, 10, "dcc");
+        psi(&mut c, 10, 0.0);
+        lease_req(&mut c, 10, SLICE);
+        c.handle(CoreEvent::Tick, Instant::now()); // grant
+        assert_eq!(n_leased(&c), 1);
+        c.handle(CoreEvent::Disconnected(10), Instant::now()); // holder cai
+        assert_eq!(n_leased(&c), 0); // lease released (DT-19)
+    }
+
+    fn reg_transport(
+        c: &mut BrokerCore,
+        sid: usize,
+        name: &str,
+        transport: TransportKind,
+    ) -> Vec<Outbound> {
+        c.handle(
+            CoreEvent::Msg(
+                sid,
+                ramshared_broker::protocol::Msg::Register {
+                    proto: PROTO_VERSION,
+                    tenant: name.into(),
+                    transport,
+                },
+            ),
+            Instant::now(),
+        )
+    }
+
+    /// SPEC windows-swap-driver DT-7: WinDrive never receives SwapOn (lease-only).
+    #[test]
+    fn windrive_does_not_receive_swap() {
+        let mut c = core(2);
+        reg_transport(&mut c, 10, "swap", TransportKind::NbdUnix);
+        reg_transport(&mut c, 20, "windrive", TransportKind::WinDrive);
+        psi(&mut c, 10, 0.0);
+        psi(&mut c, 20, 0.0);
+        let o = c.handle(CoreEvent::Tick, Instant::now());
+        let swapons: Vec<_> = o
+            .iter()
+            .filter_map(|x| match x {
+                Outbound::ToSession(sid, ramshared_broker::protocol::Msg::SwapOn { .. }) => Some(sid),
+                _ => None,
+            })
+            .collect();
+        // Only the NbdUnix tenant is in the swap round-robin; both slices go to sid 10.
+        assert!(
+            swapons.iter().all(|&sid| *sid == 10),
+            "WinDrive must not receive SwapOn; got {swapons:?}"
+        );
+        assert!(
+            !swapons.contains(&&20),
+            "WinDrive tenant must be excluded from present"
+        );
+    }
+
+    /// SPEC windows-swap-driver DT-7: WinDrive can acquire a lease (revokes swap if needed).
+    #[test]
+    fn windrive_can_lease() {
+        let mut c = core(1);
+        reg_transport(&mut c, 10, "swap", TransportKind::NbdUnix);
+        reg_transport(&mut c, 20, "windrive", TransportKind::WinDrive);
+        psi(&mut c, 10, 0.0);
+        psi(&mut c, 20, 0.0);
+        c.handle(CoreEvent::Tick, Instant::now()); // s0 → swap Active
+        assert_eq!(c.slice_map.get(0).unwrap().tenant, Some(1));
+        lease_req(&mut c, 20, SLICE);
+        let o = c.handle(CoreEvent::Tick, Instant::now());
+        assert!(
+            o.iter()
+                .any(|x| matches!(x, Outbound::ToSession(10, ramshared_broker::protocol::Msg::SwapOff { slice: 0 })))
+        );
+        c.handle(
+            CoreEvent::Msg(
+                10,
+                ramshared_broker::protocol::Msg::SwapOffDone {
+                    slice: 0,
+                    ok: true,
+                    detail: String::new(),
+                },
+            ),
+            Instant::now(),
+        );
+        c.handle(CoreEvent::ZeroDone(0, true), Instant::now());
+        let o = c.handle(CoreEvent::Tick, Instant::now());
+        assert!(
+            o.iter()
+                .any(|x| matches!(x, Outbound::ToSession(20, ramshared_broker::protocol::Msg::LeaseGranted { .. }))),
+            "WinDrive must receive LeaseGranted"
+        );
+        assert_eq!(c.slice_map.get(0).unwrap().state, SliceState::Leased);
+    }
+}
